fix diverging implicits in Natchez conversions
