+++ modules/natchez/src/main/scala/io/janstenpickle/trace4cats/natchez/conversions/NatchezToTrace4Cats.scala
@@ -1,9 +1,8 @@
 package io.janstenpickle.trace4cats.natchez.conversions
 
-import java.net.URI
-
-import _root_.natchez.{Kernel, Trace => NatchezTrace, TraceValue => V}
+import _root_.natchez.{Trace => NatchezTrace, TraceValue => V}
 import cats.Applicative
+import cats.syntax.foldable._
 import cats.syntax.functor._
 import io.janstenpickle.trace4cats.ToHeaders
 import io.janstenpickle.trace4cats.inject.Trace
@@ -20,21 +19,7 @@ import io.janstenpickle.trace4cats.model.AttributeValue.{
 import io.janstenpickle.trace4cats.model.{AttributeValue, SpanKind, SpanStatus, TraceHeaders}
 import io.janstenpickle.trace4cats.natchez.KernelConverter
 
-trait Trace4CatsConversions {
-  implicit def trace4CatsToNatchez[F[_]: Applicative](implicit trace: Trace[F]): NatchezTrace[F] =
-    new NatchezTrace[F] {
-      override def put(fields: (String, V)*): F[Unit] =
-        trace.putAll(fields.map {
-          case (k, V.StringValue(v)) => k -> StringValue(v)
-          case (k, V.NumberValue(v)) => k -> DoubleValue(v.doubleValue())
-          case (k, V.BooleanValue(v)) => k -> BooleanValue(v)
-        }: _*)
-      override def kernel: F[Kernel] = trace.headers.map(KernelConverter.to)
-      override def span[A](name: String)(k: F[A]): F[A] = trace.span[A](name)(k)
-      override def traceId: F[Option[String]] = trace.traceId
-      override def traceUri: F[Option[URI]] = Applicative[F].pure(None)
-    }
-
+trait NatchezToTrace4Cats {
   implicit def natchezToTrace4Cats[F[_]: Applicative](implicit trace: NatchezTrace[F]): Trace[F] =
     new Trace[F] {
       override def put(key: String, value: AttributeValue): F[Unit] = putAll(key -> value)
@@ -44,10 +29,10 @@ trait Trace4CatsConversions {
           case (k, DoubleValue(v)) => k -> V.NumberValue(v.value)
           case (k, LongValue(v)) => k -> V.NumberValue(v.value)
           case (k, BooleanValue(v)) => k -> V.BooleanValue(v.value)
-          case (k, StringList(v)) => k -> V.StringValue(v.toString())
-          case (k, BooleanList(v)) => k -> V.StringValue(v.toString())
-          case (k, DoubleList(v)) => k -> V.StringValue(v.toString())
-          case (k, LongList(v)) => k -> V.StringValue(v.toString())
+          case (k, StringList(v)) => k -> V.StringValue(v.value.mkString_("[", ", ", "]"))
+          case (k, BooleanList(v)) => k -> V.StringValue(v.value.mkString_("[", ", ", "]"))
+          case (k, DoubleList(v)) => k -> V.StringValue(v.value.mkString_("[", ", ", "]"))
+          case (k, LongList(v)) => k -> V.StringValue(v.value.mkString_("[", ", ", "]"))
         }: _*)
       override def span[A](name: String, kind: SpanKind)(fa: F[A]): F[A] = trace.span(name)(fa)
       override def headers(toHeaders: ToHeaders): F[TraceHeaders] = trace.kernel.map(KernelConverter.from)
