+++ modules/natchez/src/main/scala/io/janstenpickle/trace4cats/natchez/conversions/fromNatchez.scala
@@ -1,0 +1,3 @@
+package io.janstenpickle.trace4cats.natchez.conversions
+
+object fromNatchez extends NatchezToTrace4Cats
