--- modules/natchez/src/main/scala/io/janstenpickle/trace4cats/natchez/conversions/package.scala
@@ -1,3 +1,0 @@
-package io.janstenpickle.trace4cats.natchez
-
-package object conversions extends Trace4CatsConversions
