+++ modules/natchez/src/main/scala/io/janstenpickle/trace4cats/natchez/conversions/toNatchez.scala
@@ -1,0 +1,3 @@
+package io.janstenpickle.trace4cats.natchez.conversions
+
+object toNatchez extends Trace4CatsToNatchez
